# web/openapi.json
{
  "openapi" : "3.1.0",
  "info" : {
    "title" : "OpenAPI definition",
    "version" : "v0"
  },
  "paths" : {
    "/api/chapters/{code}/projects/{slug}" : {
      "get" : {
        "tags" : [ "project-controller" ],
        "operationId" : "get",
        "parameters" : [ {
          "name" : "code",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "slug",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProjectDetail"
                }
              }
            }
          }
        }
      },
      "put" : {
        "tags" : [ "project-controller" ],
        "operationId" : "update",
        "parameters" : [ {
          "name" : "code",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "slug",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/ProjectInput"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProjectDetail"
                }
              }
            }
          }
        }
      }
    },
    "/api/chapters/{code}/projects/{slug}/resources/{resourceId}" : {
      "put" : {
        "tags" : [ "project-controller" ],
        "operationId" : "updateResource",
        "parameters" : [ {
          "name" : "code",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "slug",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "resourceId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/ResourceInput"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProjectDetail"
                }
              }
            }
          }
        }
      },
      "delete" : {
        "tags" : [ "project-controller" ],
        "operationId" : "detach",
        "parameters" : [ {
          "name" : "code",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "slug",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "resourceId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProjectDetail"
                }
              }
            }
          }
        }
      }
    },
    "/api/chapters/{code}/projects/{slug}/members/{personId}" : {
      "put" : {
        "tags" : [ "project-controller" ],
        "operationId" : "setRole",
        "parameters" : [ {
          "name" : "code",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "slug",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "personId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/MemberRoleInput"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProjectDetail"
                }
              }
            }
          }
        }
      },
      "delete" : {
        "tags" : [ "project-controller" ],
        "operationId" : "removeMember",
        "parameters" : [ {
          "name" : "code",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "slug",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "personId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProjectDetail"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/chapters/{code}/sync/run" : {
      "post" : {
        "tags" : [ "project-controller" ],
        "operationId" : "runDryRun",
        "parameters" : [ {
          "name" : "code",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "202" : {
            "description" : "Accepted"
          }
        }
      }
    },
    "/api/chapters/{code}/projects" : {
      "get" : {
        "tags" : [ "project-controller" ],
        "operationId" : "list",
        "parameters" : [ {
          "name" : "code",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "$ref" : "#/components/schemas/ProjectSummary"
                  }
                }
              }
            }
          }
        }
      },
      "post" : {
        "tags" : [ "project-controller" ],
        "operationId" : "create",
        "parameters" : [ {
          "name" : "code",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/ProjectInput"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "201" : {
            "description" : "Created",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProjectDetail"
                }
              }
            }
          }
        }
      }
    },
    "/api/chapters/{code}/projects/{slug}/status" : {
      "post" : {
        "tags" : [ "project-controller" ],
        "operationId" : "status",
        "parameters" : [ {
          "name" : "code",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "slug",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/StatusChange"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProjectDetail"
                }
              }
            }
          }
        }
      }
    },
    "/api/chapters/{code}/projects/{slug}/resources" : {
      "post" : {
        "tags" : [ "project-controller" ],
        "operationId" : "attach",
        "parameters" : [ {
          "name" : "code",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "slug",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/ResourceInput"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProjectDetail"
                }
              }
            }
          }
        }
      }
    },
    "/api/chapters/{code}/projects/{slug}/members" : {
      "post" : {
        "tags" : [ "project-controller" ],
        "operationId" : "addMember",
        "parameters" : [ {
          "name" : "code",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "slug",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/MemberInput"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProjectDetail"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/chapters/{code}" : {
      "get" : {
        "tags" : [ "chapter-controller" ],
        "operationId" : "chapter",
        "parameters" : [ {
          "name" : "code",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ChapterOverview"
                }
              }
            }
          }
        }
      }
    },
    "/api/chapters/{code}/terms" : {
      "get" : {
        "tags" : [ "project-controller" ],
        "operationId" : "terms",
        "parameters" : [ {
          "name" : "code",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "$ref" : "#/components/schemas/TermOption"
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/api/chapters/{code}/resources" : {
      "get" : {
        "tags" : [ "project-controller" ],
        "operationId" : "resources",
        "parameters" : [ {
          "name" : "code",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "$ref" : "#/components/schemas/ChapterResourceOption"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/chapters/{code}/members" : {
      "get" : {
        "tags" : [ "chapter-controller" ],
        "operationId" : "members",
        "parameters" : [ {
          "name" : "code",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "$ref" : "#/components/schemas/MemberRow"
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components" : {
    "schemas" : {
      "ProjectInput" : {
        "type" : "object",
        "properties" : {
          "name" : {
            "type" : "string"
          },
          "slug" : {
            "type" : [ "string", "null" ]
          },
          "type" : {
            "type" : [ "string", "null" ]
          },
          "tags" : {
            "type" : "array",
            "items" : {
              "type" : "string"
            }
          },
          "termId" : {
            "type" : [ "string", "null" ],
            "format" : "uuid"
          },
          "partner" : {
            "type" : [ "string", "null" ]
          },
          "startsOn" : {
            "type" : [ "string", "null" ],
            "format" : "date"
          },
          "endsOn" : {
            "type" : [ "string", "null" ],
            "format" : "date"
          }
        }
      },
      "AccessPreview" : {
        "type" : "object",
        "properties" : {
          "people" : {
            "type" : "integer",
            "format" : "int32"
          },
          "grants" : {
            "type" : "integer",
            "format" : "int32"
          },
          "awaitingAgreement" : {
            "type" : "integer",
            "format" : "int32"
          },
          "asIfActive" : {
            "type" : "boolean"
          }
        }
      },
      "ActivityRow" : {
        "type" : "object",
        "properties" : {
          "at" : {
            "type" : "string",
            "format" : "date-time"
          },
          "actor" : {
            "type" : [ "string", "null" ]
          },
          "action" : {
            "type" : "string"
          },
          "detail" : {
            "type" : [ "string", "null" ]
          }
        }
      },
      "NotionPlan" : {
        "type" : "object",
        "properties" : {
          "status" : {
            "type" : "string"
          },
          "title" : {
            "type" : "string"
          },
          "pageId" : {
            "type" : [ "string", "null" ]
          },
          "parentPageId" : {
            "type" : [ "string", "null" ]
          },
          "parentPath" : {
            "type" : [ "string", "null" ]
          },
          "matchedBy" : {
            "type" : [ "string", "null" ]
          }
        }
      },
      "PlannedChange" : {
        "type" : "object",
        "properties" : {
          "kind" : {
            "type" : "string"
          },
          "tool" : {
            "type" : "string"
          },
          "resource" : {
            "type" : "string"
          },
          "person" : {
            "type" : [ "string", "null" ]
          },
          "accountId" : {
            "type" : [ "string", "null" ]
          },
          "fromAccess" : {
            "type" : [ "string", "null" ]
          },
          "toAccess" : {
            "type" : [ "string", "null" ]
          }
        }
      },
      "ProjectDetail" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "string",
            "format" : "uuid"
          },
          "name" : {
            "type" : "string"
          },
          "slug" : {
            "type" : "string"
          },
          "status" : {
            "type" : "string"
          },
          "type" : {
            "type" : [ "string", "null" ]
          },
          "tags" : {
            "type" : "array",
            "items" : {
              "type" : "string"
            }
          },
          "termId" : {
            "type" : [ "string", "null" ],
            "format" : "uuid"
          },
          "term" : {
            "type" : [ "string", "null" ]
          },
          "partner" : {
            "type" : [ "string", "null" ]
          },
          "startsOn" : {
            "type" : [ "string", "null" ],
            "format" : "date"
          },
          "endsOn" : {
            "type" : [ "string", "null" ],
            "format" : "date"
          },
          "members" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/ProjectMemberRow"
            }
          },
          "resources" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/ProjectResourceRow"
            }
          },
          "agreementsSigned" : {
            "type" : "integer",
            "format" : "int32"
          },
          "notion" : {
            "$ref" : "#/components/schemas/NotionPlan"
          },
          "preview" : {
            "$ref" : "#/components/schemas/AccessPreview"
          },
          "planned" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/PlannedChange"
            }
          },
          "lastDryRunAt" : {
            "type" : [ "string", "null" ],
            "format" : "date-time"
          },
          "plannedOutOfDate" : {
            "type" : "boolean"
          },
          "activity" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/ActivityRow"
            }
          },
          "canEdit" : {
            "type" : "boolean"
          }
        }
      },
      "ProjectMemberRow" : {
        "type" : "object",
        "properties" : {
          "personId" : {
            "type" : "string",
            "format" : "uuid"
          },
          "name" : {
            "type" : "string"
          },
          "email" : {
            "type" : [ "string", "null" ]
          },
          "status" : {
            "type" : "string"
          },
          "roleId" : {
            "type" : [ "string", "null" ],
            "format" : "uuid"
          },
          "role" : {
            "type" : [ "string", "null" ]
          },
          "agreementStatus" : {
            "type" : "string"
          },
          "lead" : {
            "type" : "boolean"
          }
        }
      },
      "ProjectResourceRow" : {
        "type" : "object",
        "properties" : {
          "resourceId" : {
            "type" : "string",
            "format" : "uuid"
          },
          "tool" : {
            "type" : "string"
          },
          "name" : {
            "type" : "string"
          },
          "state" : {
            "type" : "string"
          },
          "audience" : {
            "type" : "string"
          },
          "roleId" : {
            "type" : [ "string", "null" ],
            "format" : "uuid"
          },
          "role" : {
            "type" : [ "string", "null" ]
          },
          "access" : {
            "type" : "string"
          },
          "requiresAgreement" : {
            "type" : "boolean"
          },
          "tags" : {
            "type" : "array",
            "items" : {
              "type" : "string"
            }
          },
          "sharedWith" : {
            "type" : "array",
            "items" : {
              "type" : "string"
            }
          },
          "existsInTool" : {
            "type" : "boolean"
          },
          "getting" : {
            "type" : "integer",
            "format" : "int32"
          },
          "awaitingAgreement" : {
            "type" : "integer",
            "format" : "int32"
          }
        }
      },
      "ResourceInput" : {
        "type" : "object",
        "properties" : {
          "resourceId" : {
            "type" : [ "string", "null" ],
            "format" : "uuid"
          },
          "tool" : {
            "type" : [ "string", "null" ]
          },
          "name" : {
            "type" : [ "string", "null" ]
          },
          "audience" : {
            "type" : "string"
          },
          "roleId" : {
            "type" : [ "string", "null" ],
            "format" : "uuid"
          },
          "access" : {
            "type" : "string"
          },
          "requiresAgreement" : {
            "type" : "boolean"
          },
          "tags" : {
            "type" : "array",
            "items" : {
              "type" : "string"
            }
          }
        }
      },
      "MemberRoleInput" : {
        "type" : "object",
        "properties" : {
          "roleId" : {
            "type" : [ "string", "null" ],
            "format" : "uuid"
          }
        }
      },
...
      "StatusChange" : {
        "type" : "object",
        "properties" : {
          "status" : {
            "type" : "string"
          }
        }
      },
      "MemberInput" : {
        "type" : "object",
        "properties" : {
          "personId" : {
            "type" : "string",
            "format" : "uuid"
          },
          "roleId" : {
            "type" : [ "string", "null" ],
            "format" : "uuid"
          }
        }
      },
...
      "ChapterOverview" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "string",
            "format" : "uuid"
          },
          "code" : {
            "type" : "string"
          },
          "name" : {
            "type" : "string"
          },
          "status" : {
            "type" : "string"
          },
          "role" : {
            "type" : "string"
          },
          "stats" : {
            "$ref" : "#/components/schemas/ChapterStats"
          },
          "registrationLink" : {
            "type" : "string"
          },
          "notionProblems" : {
            "type" : "integer",
            "format" : "int32"
          }
        }
      },
      "ChapterStats" : {
        "type" : "object",
        "properties" : {
          "activeMembers" : {
            "type" : "integer",
            "format" : "int32"
          },
          "alumni" : {
            "type" : "integer",
            "format" : "int32"
          },
          "liveProjects" : {
            "type" : "integer",
            "format" : "int32"
          },
          "leads" : {
            "type" : "integer",
            "format" : "int32"
          }
        }
      },
      "TermOption" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "string",
            "format" : "uuid"
          },
          "label" : {
            "type" : "string"
          }
        }
      },
...
      "ChapterResourceOption" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "string",
            "format" : "uuid"
          },
          "tool" : {
            "type" : "string"
          },
          "name" : {
            "type" : "string"
          },
          "state" : {
            "type" : "string"
          },
          "projects" : {
            "type" : "array",
            "items" : {
              "type" : "string"
            }
          }
        }
      },
      "ProjectSummary" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "string",
            "format" : "uuid"
          },
          "name" : {
            "type" : "string"
          },
          "slug" : {
            "type" : "string"
          },
          "status" : {
            "type" : "string"
          },
          "type" : {
            "type" : [ "string", "null" ]
          },
          "tags" : {
            "type" : "array",
            "items" : {
              "type" : "string"
            }
          },
          "term" : {
            "type" : [ "string", "null" ]
          },
          "partner" : {
            "type" : [ "string", "null" ]
          },
          "members" : {
            "type" : "integer",
            "format" : "int32"
          },
          "resources" : {
            "type" : "integer",
            "format" : "int32"
          },
          "updatedAt" : {
            "type" : "string",
            "format" : "date-time"
          }
        }
      },
      "AccountSummary" : {
        "type" : "object",
        "properties" : {
          "tool" : {
            "type" : "string"
          },
          "state" : {
            "type" : "string"
          }
        }
      },
      "MemberRow" : {
        "type" : "object",
        "properties" : {
          "personId" : {
            "type" : "string",
            "format" : "uuid"
          },
          "name" : {
            "type" : "string"
          },
          "email" : {
            "type" : [ "string", "null" ]
          },
          "status" : {
            "type" : "string"
          },
          "kind" : {
            "type" : "string"
          },
          "chapterRole" : {
            "type" : [ "string", "null" ]
          },
          "title" : {
            "type" : [ "string", "null" ]
          },
          "projects" : {
            "type" : "array",
            "items" : {
              "type" : "string"
            }
          },
          "accounts" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/AccountSummary"
            }
          },
          "claimed" : {
            "type" : "boolean"
          }
        }
      },
...
    }
  }
}